images/sync/requirements.txt
flask==0.12.1
gitpython==2.1.1
